+++ eventpass/eventpass_backend/eventpass_api/models.py
@@ -1,66 +1,66 @@
 from django.db import models
-from django.contrib.auth.models import AbstractUser
+from django.contrib.auth.models import AbstractUser, BaseUserManager
 
 
-class Venue(models.Model):
-    name = models.CharField(max_length=255)
-    address = models.CharField(max_length=255)
-    city = models.CharField(max_length=255)
-    state = models.CharField(max_length=150)
-    zip_code = models.IntegerField()
-    capacity = models.IntegerField()
-    indoor = models.BooleanField()
-    outdoor = models.BooleanField()
-    photo_url = models.CharField(max_length=500, default='a string')
+# class Venue(models.Model):
+#     name = models.CharField(max_length=255)
+#     address = models.CharField(max_length=255)
+#     city = models.CharField(max_length=255)
+#     state = models.CharField(max_length=150)
+#     zip_code = models.IntegerField()
+#     capacity = models.IntegerField()
+#     indoor = models.BooleanField()
+#     outdoor = models.BooleanField()
+#     photo_url = models.CharField(max_length=500, default='a string')
 
-    def __str__(self):
-        return self.name
-
-
-class Ticket(models.Model):
-    event = models.ForeignKey(
-        Event, on_delete=models.CASCADE, related_name='tickets')
-    customer = models.ForeignKey(
-        Customer, on_delete=models.CASCADE, related_name='tickets')
-    price = models.DecimalField(max_digits=5, decimal_places=2)
-    seating = models.CharField(max_length=150)
-    number_of_tickets = models.IntegerField()
-    tickets_sold = models.IntegerField()
-
-    def __str__(self):
-        return self.price
+#     def __str__(self):
+#         return self.name
 
 
-class Customer(models.Model):
-    event = models.ForeignKey(
-        Event, on_delete=models.CASCADE, related_name='customers')
-    ticket = models.ForeignKey(
-        Ticket, on_delete=models.CASCADE, related_name='customers')
-    name = models.CharField(max_length=255)
-    address = models.CharField(max_length=255)
-    city = models.CharField(max_length=255)
-    state = models.CharField(max_length=150)
-    zip_code = models.IntegerField()
-    purchases = models.CharField(max_length=500)
-    photo_url = models.CharField(max_length=500, default='a string')
+# class Ticket(models.Model):
+#     event = models.ForeignKey(
+#         Event, on_delete=models.CASCADE, related_name='tickets')
+#     customer = models.ForeignKey(
+#         Customer, on_delete=models.CASCADE, related_name='tickets')
+#     price = models.DecimalField(max_digits=5, decimal_places=2)
+#     seating = models.CharField(max_length=150)
+#     number_of_tickets = models.IntegerField()
+#     tickets_sold = models.IntegerField()
 
-    def __str__(self):
-        return self.name
+#     def __str__(self):
+#         return self.price
 
 
-class Event(models.Model):
-    venue = models.ForeignKey(
-        Venue, on_delete=models.CASCADE, related_name='events')
-    customer = models.ForeignKey(
-        Customer, on_delete=models.CASCADE, related_name='events')
-    ticket = models.ForeignKey(
-        Ticket, on_delete=models.CASCADE, related_name='events')
-    name = models.CharField(max_length=500)
-    datetime = models.DateTimeField()
-    photo_url = models.CharField(max_length=500, default='a string')
+# class Customer(models.Model):
+#     event = models.ForeignKey(
+#         Event, on_delete=models.CASCADE, related_name='customers')
+#     ticket = models.ForeignKey(
+#         Ticket, on_delete=models.CASCADE, related_name='customers')
+#     name = models.CharField(max_length=255)
+#     address = models.CharField(max_length=255)
+#     city = models.CharField(max_length=255)
+#     state = models.CharField(max_length=150)
+#     zip_code = models.IntegerField()
+#     purchases = models.CharField(max_length=500)
+#     photo_url = models.CharField(max_length=500, default='a string')
 
-    def __str__(self):
-        return self.name
+#     def __str__(self):
+#         return self.name
+
+
+# class Event(models.Model):
+#     venue = models.ForeignKey(
+#         Venue, on_delete=models.CASCADE, related_name='events')
+#     customer = models.ForeignKey(
+#         Customer, on_delete=models.CASCADE, related_name='events')
+#     ticket = models.ForeignKey(
+#         Ticket, on_delete=models.CASCADE, related_name='events')
+#     name = models.CharField(max_length=500)
+#     datetime = models.DateTimeField()
+#     photo_url = models.CharField(max_length=500, default='a string')
+
+#     def __str__(self):
+#         return self.name
 
 
 class User(AbstractUser):
@@ -73,22 +73,22 @@ class User(AbstractUser):
     REQUIRED_FIELDS = []
 
 
-# class UserManager(BaseUserManager):
-#     def create_user(self, email, name, password=None, **extra_fields):
-#         if not email:
-#             raise ValueError('The Email field must be set')
-#         email = self.normalize_email(email)
-#         user = self.model(email=email, name=name, **extra_fields)
-#         user.set_password(password)
-#         user.save(using=self._db)
-#         return user
+class UserManager(BaseUserManager):
+    def create_user(self, email, name, password=None, **extra_fields):
+        if not email:
+            raise ValueError('The Email field must be set')
+        email = self.normalize_email(email)
+        user = self.model(email=email, name=name, **extra_fields)
+        user.set_password(password)
+        user.save(using=self._db)
+        return user
 
-#     def create_superuser(self, email, name, password=None, **extra_fields):
-#         extra_fields.setdefault('is_staff', True)
-#         extra_fields.setdefault('is_superuser', True)
-#         return self.create_user(email, name, password, **extra_fields)
+    def create_superuser(self, email, name, password=None, **extra_fields):
+        extra_fields.setdefault('is_staff', True)
+        extra_fields.setdefault('is_superuser', True)
+        return self.create_user(email, name, password, **extra_fields)
 
-#     REQUIRED_FIELDS = ['name']
+    REQUIRED_FIELDS = ['name']
 
 
-# objects = UserManager()
+objects = UserManager()
