+++ eventpass/eventpass_backend/eventpass_django/settings.py
